Make the ingester much more fallible.
This set of changes makes the ingester fail if symbolication or
signature generation encounter any errors. It's important for these to
be louder than a log line, because the data we ingest is persisted and
stored as a source of truth; as such, we don't want partial data
resulting from errors that were ignored in the overall program. If there
are some errors that we deem acceptable or even expected, we can account
for those specifically.

This also adds a retry mechanism to breakpad file fetching. We don't yet
add it to symsrv (we'd have to register a SymsrvObserver and check
download failures), however top-level symsrv errors now result in
program termination.

There is a new `keep_going` configuration value to not fail and just
emit partial information. Right now it's all-or-nothing as far as
symbolication is concerned: e.g., a server failure to fetch a symbol
file will result in symbolication for the entire ping to fail, even if
other frames with other symbol file sources would have succeeded.