Bug 1959459 - Store (when downloading) and upload the ingestion partitioning configuration.
This adds a new table (`ingest_config`) which will store how we
partitioned the data when downloading. This is useful for determining
how representative the sampled data is.

The upload script has been changed to run all uploads as a single
transaction, and to _DELETE_ existing data for the upload date so that
all data in the tables represents a single download/process/upload run.